refactoring Value to use stack allocated values for emp objects

+++ src/main/cpp/query_table/query_tuple.cpp
@@ -143,7 +143,7 @@ QueryTuple& QueryTuple::operator=(const QueryTuple& src) {
     fields_.resize(src.getFieldCount());
 
 
-    for(int i = 0; i < src.getFieldCount(); ++i) {
+    for(size_t i = 0; i < src.getFieldCount(); ++i) {
         fields_[i] = src.fields_[i];
     }
